Composite key tests collected in one test class.
