+++ src/test/java/com/zaxxer/sansorm/SqlClosureElfTest.java
@@ -8,6 +8,7 @@ import org.sansorm.TestUtils;
 import javax.persistence.Column;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.sql.Connection;
 import java.sql.SQLException;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -34,55 +35,4 @@ public class SqlClosureElfTest {
       assertThat(SqlClosureElf.getInClausePlaceholdersForCount(5)).isEqualTo(" (?,?,?,?,?) ");
       assertThatIllegalArgumentException().isThrownBy(() -> SqlClosureElf.getInClausePlaceholdersForCount(-1));
    }
-
-   @Table
-   public static class TestClass2 {
-      @Id
-      String id1 = "id1";
-      @Id
-      String id2 = "id2";
-      @Column
-      String field;
-   }
-
-   /**
-    * TODO Composite primary key is not supported???
-    * <p>
-    * java.lang.RuntimeException: org.h2.jdbc.JdbcSQLException: NULL not allowed for column "ID1"; SQL statement: INSERT INTO TestClass2(field) VALUES (?)
-    */
-   @Test
-   public void insertObjectCompositeKeyH2() throws SQLException {
-
-      JdbcDataSource ds = TestUtils.makeH2DataSource();
-      SansOrm.initializeTxNone(ds);
-      try {
-         SqlClosureElf.executeUpdate(
-            " CREATE TABLE TestClass2 ("
-               + "id1 VARCHAR(128) NOT NULL, "
-               + "id2 VARCHAR(128) NOT NULL, "
-               + "field VARCHAR(128), "
-               + "PRIMARY KEY (id1, id2)"
-               + ")");
-
-         String id1 = "id1";
-         String id2 = "id2";
-         String field = "field";
-
-         SqlClosureElfTest.TestClass2 obj = SqlClosureElf.insertObject(new SqlClosureElfTest.TestClass2());
-         assertEquals(id1, obj.id1);
-         obj = SqlClosureElf.getObjectById(SqlClosureElfTest.TestClass2.class, obj.id1, obj.id2);
-         assertNotNull(obj);
-
-         SqlClosureElf.executeUpdate("update TestClass2 set field = 'changed'");
-
-         SqlClosureElfTest.TestClass2 obj2 = OrmElf.refresh(ds.getConnection(), obj);
-         assertTrue(obj == obj2);
-         assertEquals("changed", obj.field);
-
-      }
-      finally {
-         SqlClosureElf.executeUpdate("DROP TABLE TestClass2");
-      }
-   }
-
 }
